+++ changelog.md
@@ -37,6 +37,7 @@
 * #1295 Sales tracking (tracking pixel) for Billiger.de
 * XML and CSV export of shopping cart and wishlist items
 * #1363 Make storing of IP addresses optional
+* #729 Option for automatic order amount capturing when the shipping status changed to "shipped"
 
 ### Improvements
 * Target .NET Framework changed: 4.5.2 > 4.6.1.